+++ changelog.md
@@ -1,3 +1,8 @@
 # Icon Ninja Changelog
-## 1.0.0
+## v1.0.1
+- Added margin of +/-5px on icons
+- Changed intersection algorithm
+- Fixed bug that made the player unable to jump when playing a level
+- Fixed bug that didnt hide controller glyphs or button hints when going into gameplay
+## v1.0.0
 - Initial Release!